tests/e2e/requirements.txt
playwright>=1.44.0
firebase-admin>=6.5.0
gspread>=6.1.0
google-auth>=2.29.0
google-auth-oauthlib>=1.2.0
requests>=2.31.0
python-dotenv>=1.0.0